+++ test/runtests.jl
@@ -14,13 +14,20 @@ using Test
     @test alm_index(triangular_alm, 5, 4) == 39
     alm = make_alm_info(10, 10, 1, 0:10)
     @test alm_index(alm, 5, 4) == 9
+    general_alm = make_general_alm_info(3, 4, 1, 0:3, [0, 3, 5, 6])
+    @test alm_index(general_alm, 3, 2) == 8
+    alm = make_mmajor_complex_alm_info(6, 1, nothing)
+    @test alm_index(alm, 5, 4) == 23
 
     # test total counts
     alm = make_alm_info(10, 10, 1, 0:10)
     @test alm_count(alm) == 66
     triangular_alm = make_triangular_alm_info(10, 10, 1)
     @test alm_count(triangular_alm) == 66
-
+    general_alm = make_general_alm_info(3, 4, 1, 0:3, [0, 3, 5, 6])
+    @test alm_count(general_alm) == 10
+    alm = make_mmajor_complex_alm_info(6, 1, nothing)
+    @test alm_count(alm) == 28
 end
 
 ## tests relating to pixelization properties
@@ -34,9 +41,17 @@ end
     # test the total number of pixels
     geom_info = make_healpix_geom_info(16, 1)
     @test map_size(geom_info) == 3072
-    geom_info = make_weighted_healpix_geom_info(
-        16, 1, ones(16 * 4 - 1))
+    geom_info = make_weighted_healpix_geom_info(16, 1, ones(16 * 4 - 1))
     @test map_size(geom_info) == 3072
+
+    #test the number of pixels in a subset
+    geom_info = make_subset_healpix_geom_info(8, 1, 6, [14, 18, 9, 23, 4, 28])
+    @test map_size(geom_info) == 160
+
+    #and finalizer
+    @test geom_info.ptr != C_NULL
+    Libsharp.destroy_geom_info(geom_info)
+    @test geom_info.ptr == C_NULL
 end
 
 ## tests relating to spherical harmonic transform map2alm